refactor: break down index.js monolith into modular files (#2770)
Extracts inline routes, middlewares, and repositories from the 1700-line index.js file into their respective modules to improve maintainability and separation of concerns.
- Created contentStore and refactored auth middlewares
- Moved inline notifications and portfolio routes to route files
- Created circuit breaker route file
- Extracted multer upload logic to middleware

Fixes #2747
