+++ server/controllers/coreTeamController.js
@@ -31,6 +31,26 @@ export const adminListCoreTeamMembers = wrapAsync(async (req, res) => {
   return res.json({ members });
 });
 
+export const publicListMembers = wrapAsync(async (req, res) => {
+  try {
+    const rawMembers = await coreTeamService.listMembers();
+    const members = (rawMembers || []).map((m) => {
+      let email = m.email || null;
+      if (email && !email.toLowerCase().endsWith('@glbajajgroup.org')) {
+        email = null;
+      }
+      return {
+        ...m,
+        email,
+        whatsapp: 'https://chat.whatsapp.com/FhpJEaod2g419jFMfqrhGZ',
+      };
+    });
+    return res.json({ members });
+  } catch (e) {
+    return res.status(500).json({ error: e?.message || 'Failed to load core team' });
+  }
+});
+
 export const adminAddCoreTeamMember = wrapAsync(async (req, res) => {
   const body = req.body || {};
   const member = {
