+++ server/index.js
@@ -112,6 +112,33 @@ import feedbackRouter from './routes/feedbackRoutes.js';
 import * as slackController from './controllers/slackController.js';
 import { startStreamingWorkers } from './streaming/startStreamingWorkers.js';
 
+import {
+  listEventsStore,
+  createEventStore,
+  updateEventStore,
+  deleteEventStore,
+  listActivityEventsStore,
+  createActivityEventStore,
+  deleteActivityEventStore,
+  listCoreTeamStore,
+  createCoreTeamStore,
+  deleteCoreTeamStore,
+  appendToSupabaseForms,
+  timingSafeStringEqual,
+  toSafeString,
+  validateWhatsApp,
+  validateSection,
+  sanitizeEvent,
+  normalizePhone
+} from './repositories/contentStore.js';
+
+import { checkPasskeyLockout, recordFailedPasskeyAttempt, clearPasskeyAttempts } from './middleware/auth/passkeyLockout.js';
+import { checkActivityAuthLockout, recordFailedActivityAuth, clearActivityAuthAttempts, canManageActivityEvent } from './middleware/auth/activityAuth.js';
+import { requireNotificationPrefAuth, requireMentorshipAuth } from './middleware/auth/customAuth.js';
+import { uploadWithMagicCheck, validateMagicBytes, UPLOADS_DIR } from './middleware/uploadMiddleware.js';
+
+import circuitBreakerRouter from './routes/circuitBreaker.js';
+
 validateLimiters();
 
 const __filename = fileURLToPath(import.meta.url);
@@ -449,747 +476,7 @@ const ADMIN_EVENT_PASSWORD = requiredStrongPassword('ADMIN_EVENT_PASSWORD');
 const SESSION_SECRET = requiredStrongPassword('SESSION_SECRET');
 
 // ── File Upload Configuration ──
-const UPLOADS_DIR = path.join(__dirname, 'uploads');
-try {
-  fs.mkdirSync(UPLOADS_DIR, { recursive: true });
-} catch (_) {}
-
-const storage = multer.diskStorage({
-  destination: (_req, _file, cb) => cb(null, UPLOADS_DIR),
-  filename: (_req, file, cb) => {
-    const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1e9);
-    const ext = path.extname(file.originalname) || '';
-    cb(null, `${uniqueSuffix}${ext}`);
-  },
-});
-
-const MAGIC_BYTES = {
-  'application/pdf': [[0x25, 0x50, 0x44, 0x46]],
-  'image/png': [[0x89, 0x50, 0x4e, 0x47]],
-  'image/jpeg': [[0xff, 0xd8, 0xff]],
-  'image/gif': [[0x47, 0x49, 0x46]],
-  'image/webp': [[0x52, 0x49, 0x46, 0x46]],
-  'application/zip': [
-    [0x50, 0x4b, 0x03, 0x04],
-    [0x50, 0x4b, 0x05, 0x06],
-    [0x50, 0x4b, 0x07, 0x08],
-  ],
-  'application/x-zip-compressed': [[0x50, 0x4b, 0x03, 0x04]],
-  'application/vnd.openxmlformats-officedocument.wordprocessingml.document': [
-    [0x50, 0x4b, 0x03, 0x04],
-  ],
-  'application/vnd.openxmlformats-officedocument.presentationml.presentation': [
-    [0x50, 0x4b, 0x03, 0x04],
-  ],
-  'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet': [[0x50, 0x4b, 0x03, 0x04]],
-  'text/plain': [],
-  'text/markdown': [],
-  'application/json': [],
-};
-
-function validateMagicBytes(filepath, mimeType) {
-  const signatures = MAGIC_BYTES[mimeType];
-  if (!signatures || signatures.length === 0) return true;
-  const fd = fs.openSync(filepath, 'r');
-  const buffer = Buffer.alloc(8);
-  fs.readSync(fd, buffer, 0, 8, 0);
-  fs.closeSync(fd);
-  return signatures.some((sig) => sig.every((byte, i) => buffer[i] === byte));
-}
-
-const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB
-
-const fileFilter = (_req, file, cb) => {
-  if (!file.originalname || file.originalname.includes('..') || file.originalname.includes('/')) {
-    return cb(new Error('Invalid file name'), false);
-  }
-  const allowedMimes = Object.keys(MAGIC_BYTES);
-  if (allowedMimes.includes(file.mimetype)) {
-    cb(null, true);
-  } else {
-    cb(new Error(`File type ${file.mimetype} not allowed`), false);
-  }
-};
-
-const upload = multer({
-  storage,
-  fileFilter,
-  limits: { fileSize: MAX_FILE_SIZE },
-});
-
-const uploadWithMagicCheck = (req, res, next) => {
-  upload.single('file')(req, res, (err) => {
-    if (err) {
-      if (err instanceof multer.MulterError && err.code === 'LIMIT_FILE_SIZE') {
-        return res.status(413).json({ error: 'File too large. Maximum size is 10MB.' });
-      }
-      return res.status(400).json({ error: err.message });
-    }
-    if (req.file && !validateMagicBytes(req.file.path, req.file.mimetype)) {
-      fs.unlink(req.file.path, () => {});
-      return res.status(400).json({ error: 'File content does not match its declared type.' });
-    }
-    next();
-  });
-};
-
-// Serve uploaded files statically
 app.use('/uploads', express.static(UPLOADS_DIR));
-
-getPublicAppUrl();
-
-async function ensureContentFile() {
-  const dir = path.dirname(CONTENT_FILE);
-  await fsp.mkdir(dir, { recursive: true });
-  try {
-    await fsp.access(CONTENT_FILE);
-  } catch {
-    await fsp.writeFile(CONTENT_FILE, JSON.stringify(defaultContent, null, 2), 'utf8');
-  }
-}
-const fileMutex = new Mutex();
-
-export async function runWithFileLock(callback) {
-  return await fileMutex.runExclusive(callback);
-}
-
-async function readContent() {
-  await ensureContentFile();
-  const raw = await fsp.readFile(CONTENT_FILE, 'utf8');
-  return JSON.parse(raw);
-}
-
-async function writeContent(content) {
-  await ensureContentFile();
-  await fsp.writeFile(CONTENT_FILE, JSON.stringify(content, null, 2), 'utf8');
-}
-
-let contentLock = Promise.resolve();
-
-function withContentLock(fn) {
-  let release;
-  const next = new Promise((resolve) => {
-    release = resolve;
-  });
-  const current = contentLock;
-  contentLock = next;
-  return current.then(() => fn()).finally(() => release());
-}
-
-const _rawSupabaseRequest = async function _rawSupabaseRequest(
-  pathname,
-  { method = 'GET', body } = {}
-) {
-  if (!HAS_SUPABASE) throw new Error('Supabase is not configured');
-  const res = await tracedFetch(`${SUPABASE_URL}/rest/v1/${pathname}`, {
-    method,
-    headers: {
-      apikey: SUPABASE_SERVICE_KEY,
-      Authorization: `Bearer ${SUPABASE_SERVICE_KEY}`,
-      'Content-Type': 'application/json',
-      Prefer: method === 'GET' ? 'count=exact' : 'return=representation',
-    },
-    body: body ? JSON.stringify(body) : undefined,
-  });
-
-  if (!res.ok) {
-    const text = await res.text();
-    throw new Error(`Supabase error (${res.status}): ${text}`);
-  }
-  const text = await res.text();
-  return text ? JSON.parse(text) : [];
-};
-
-export { _rawSupabaseRequest as supabaseRequest };
-
-export const supabaseBreaker = circuitBreakerRegistry.register(
-  'index-supabase',
-  new CircuitBreaker(_rawSupabaseRequest, {
-    name: 'index-supabase',
-    failureThreshold: 5,
-    successThreshold: 2,
-    coolDownPeriod: 10000,
-    maxCoolDownPeriod: 60000,
-  })
-);
-
-// Paginated variant: appends LIMIT/OFFSET to a PostgREST GET request and reads
-// the total row count from the Content-Range response header (sent when
-// Prefer: count=exact is set). Returns { rows, total } instead of a bare array.
-async function supabasePaginatedRequest(pathname, page, limit) {
-  if (!HAS_SUPABASE) throw new Error('Supabase is not configured');
-  const offset = (page - 1) * limit;
-  const separator = pathname.includes('?') ? '&' : '?';
-  const url = `${SUPABASE_URL}/rest/v1/${pathname}${separator}limit=${limit}&offset=${offset}`;
-  const res = await tracedFetch(url, {
-    method: 'GET',
-    headers: {
-      apikey: SUPABASE_SERVICE_KEY,
-      Authorization: `Bearer ${SUPABASE_SERVICE_KEY}`,
-      'Content-Type': 'application/json',
-      Prefer: 'count=exact',
-    },
-  });
-  if (!res.ok) {
-    const text = await res.text();
-    throw new Error(`Supabase error (${res.status}): ${text}`);
-  }
-  const text = await res.text();
-  const rows = text ? JSON.parse(text) : [];
-  // Content-Range format from PostgREST: "0-19/150" or "*/0" when empty
-  const contentRange = res.headers.get('content-range') || '';
-  const totalMatch = contentRange.match(/\/(\d+)$/);
-  const total = totalMatch ? parseInt(totalMatch[1], 10) : rows.length;
-  return { rows, total };
-}
-
-// Parses ?page and ?limit from a request query object, clamps to safe bounds,
-// and returns normalised integers. Defaults: page=1, limit=20, cap=100.
-function parsePagination(query) {
-  const page = Math.max(1, parseInt(query.page, 10) || 1);
-  const limit = Math.min(100, Math.max(1, parseInt(query.limit, 10) || 20));
-  return { page, limit };
-}
-
-function toSafeString(value, max = 4000) {
-  return String(value ?? '')
-    .trim()
-    .slice(0, max);
-}
-
-function validateWhatsApp(str) {
-  const v = String(str || '').trim();
-  if (!/^\d{10}$/.test(v)) throw new Error('WhatsApp must be exactly 10 digits');
-  return v;
-}
-
-function validateSection(str) {
-  const v = String(str || '')
-    .trim()
-    .toUpperCase();
-  if (!/^[A-Z]$/.test(v)) throw new Error('Section must be a single letter (A-Z)');
-  return v;
-}
-
-function sanitizeEvent(input = {}) {
-  const status = input.status === 'upcoming' ? 'upcoming' : 'completed';
-  const tags = Array.isArray(input.tags)
-    ? input.tags
-        .map((t) => toSafeString(t, 40))
-        .filter(Boolean)
-        .slice(0, 12)
-    : String(input.tags || '')
-        .split(',')
-        .map((t) => t.trim())
-        .filter(Boolean)
-        .slice(0, 12);
-
-  return {
-    id:
-      toSafeString(input.id || input.shortName || input.name, 80)
-        .toLowerCase()
-        .replace(/[^a-z0-9]+/g, '-')
-        .replace(/^-+|-+$/g, '') || `event-${Date.now()}`,
-    name: toSafeString(input.name, 120),
-    shortName: toSafeString(input.shortName || input.name, 60),
-    date: toSafeString(input.date, 80),
-    description: toSafeString(input.description, 1200),
-    status,
-    icon: toSafeString(input.icon || 'Pin', 32),
-    tags,
-  };
-}
-
-function normalizePhone(value) {
-  return String(value || '').replace(/[^\d]/g, '');
-}
-
-async function canManageActivityEvent({ name, email, phone, password }) {
-  const expectedPassword = process.env.ADMIN_EVENT_PASSWORD;
-  // Use constant-time comparison to prevent timing-based password recovery.
-  if (!timingSafeStringEqual(String(password ?? ''), expectedPassword)) {
-    return false;
-  }
-  const n = String(name || '')
-    .trim()
-    .toLowerCase();
-  const e = String(email || '')
-    .trim()
-    .toLowerCase();
-  const p = normalizePhone(phone);
-
-  const members = await listCoreTeamStore();
-  return members.some(
-    (m) =>
-      m.name.toLowerCase() === n && m.email.toLowerCase() === e && normalizePhone(m.whatsapp) === p
-  );
-}
-
-async function listEventsStore({ page = 1, limit = 20 } = {}) {
-  if (HAS_SUPABASE) {
-    const { rows, total } = await supabasePaginatedRequest(
-      'events?select=*&order=created_at.desc',
-      page,
-      limit
-    );
-    return {
-      events: rows.map((r) =>
-        sanitizeEventRecord({
-          id: r.id,
-          name: r.name,
-          shortName: r.short_name || r.shortName || r.name,
-          date: r.date_text || r.date,
-          description: r.description,
-          status: r.status,
-          icon: r.icon || 'Pin',
-          tags: Array.isArray(r.tags) ? r.tags : [],
-          createdAt: r.created_at,
-          updatedAt: r.updated_at,
-        })
-      ),
-      total,
-    };
-  }
-  const content = await readContent();
-  const all = (content.events || []).map((event) => sanitizeEventRecord(event));
-  const total = all.length;
-  const start = (page - 1) * limit;
-  return { events: all.slice(start, start + limit), total };
-}
-
-const ALLOWED_EVENT_FIELDS = [
-  'id',
-  'name',
-  'description',
-  'date_text',
-  'time',
-  'location',
-  'type',
-  'mode',
-  'category',
-  'tags',
-  'image_url',
-  'registration_link',
-  'capacity',
-  'registered_count',
-  'price',
-  'created_at',
-  'updated_at',
-];
-
-function sanitizeEventRecord(event) {
-  if (!event || typeof event !== 'object') return event;
-  const sanitized = {};
-  for (const field of ALLOWED_EVENT_FIELDS) {
-    if (field in event) sanitized[field] = event[field];
-  }
-  return sanitized;
-}
-
-async function createEventStore(event) {
-  if (HAS_SUPABASE) {
-    let payload = {
-      id: event.id,
-      name: event.name,
-      short_name: event.shortName,
-      date_text: event.date,
-      description: event.description,
-      status: event.status,
-      icon: event.icon,
-      tags: event.tags,
-    };
-
-    let row;
-    try {
-      [row] = await supabaseRequest('events', {
-        method: 'POST',
-        body: [payload],
-      });
-    } catch (e) {
-      // Retry with suffix if id collision occurs.
-      payload = { ...payload, id: `${event.id}-${Date.now()}` };
-      [row] = await supabaseRequest('events', {
-        method: 'POST',
-        body: [payload],
-      });
-    }
-    return sanitizeEventRecord({
-      id: row.id,
-      name: row.name,
-      shortName: row.short_name || row.name,
-      date: row.date_text,
-      description: row.description,
-      status: row.status,
-      icon: row.icon || 'Pin',
-      tags: Array.isArray(row.tags) ? row.tags : [],
-      createdAt: row.created_at,
-      updatedAt: row.updated_at,
-    });
-  }
-
-  // Safe atomic fallback operation preventing data loss using async-mutex
-  return withContentLock(async () => {
-    const content = await readContent();
-    content.events.unshift({
-      ...event,
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-    });
-    await writeContent(content);
-    return sanitizeEventRecord(content.events[0]);
-  });
-}
-async function updateEventStore(id, patch) {
-  if (HAS_SUPABASE) {
-    const [row] = await supabaseRequest(`events?id=eq.${encodeURIComponent(id)}`, {
-      method: 'PATCH',
-      body: {
-        name: patch.name,
-        short_name: patch.shortName,
-        date_text: patch.date,
-        description: patch.description,
-        status: patch.status,
-        icon: patch.icon,
-        tags: patch.tags,
-        updated_at: new Date().toISOString(),
-      },
-    });
-    if (!row) return null;
-    return sanitizeEventRecord({
-      id: row.id,
-      name: row.name,
-      shortName: row.short_name || row.name,
-      date: row.date_text,
-      description: row.description,
-      status: row.status,
-      icon: row.icon || 'Pin',
-      tags: Array.isArray(row.tags) ? row.tags : [],
-      createdAt: row.created_at,
-      updatedAt: row.updated_at,
-    });
-  }
-  return withContentLock(async () => {
-    const content = await readContent();
-    const idx = content.events.findIndex((e) => e.id === id);
-    if (idx < 0) return null;
-    content.events[idx] = {
-      ...content.events[idx],
-      ...patch,
-      id,
-      updatedAt: new Date().toISOString(),
-    };
-    await writeContent(content);
-    return sanitizeEventRecord(content.events[idx]);
-  });
-}
-
-async function deleteEventStore(id) {
-  if (HAS_SUPABASE) {
-    const rows = await supabaseRequest(`events?id=eq.${encodeURIComponent(id)}`, {
-      method: 'DELETE',
-    });
-    return Array.isArray(rows) && rows.length > 0;
-  }
-  return withContentLock(async () => {
-    const content = await readContent();
-    const before = content.events.length;
-    content.events = content.events.filter((e) => e.id !== id);
-    if (content.events.length === before) return false;
-    await writeContent(content);
-    return true;
-  });
-}
-
-async function listActivityEventsStore(activityKey, { page = 1, limit = 20 } = {}) {
-  if (HAS_SUPABASE) {
-    const { rows, total } = await supabasePaginatedRequest(
-      `activity_events?activity_key=eq.${encodeURIComponent(activityKey)}&select=*&order=created_at.desc`,
-      page,
-      limit
-    );
-    return {
-      events: rows.map((r) =>
-        sanitizeActivityEventRecord({
-          id: r.id,
-          name: r.name,
-          date: r.date_text || r.date,
-          tagline: r.tagline,
-          description: r.description,
-          status: r.status || 'completed',
-          createdAt: r.created_at,
-        })
-      ),
-      total,
-    };
-  }
-  const content = await readContent();
-  const all = (content.activityEvents?.[activityKey] || []).map((event) =>
-    sanitizeActivityEventRecord(event)
-  );
-  const total = all.length;
-  const start = (page - 1) * limit;
-  return { events: all.slice(start, start + limit), total };
-}
-
-function sanitizeActivityEventRecord(event) {
-  if (!event || typeof event !== 'object') return event;
-  const { createdBy, ...safe } = event;
-  return safe;
-}
-
-async function createActivityEventStore(activityKey, event) {
-  if (HAS_SUPABASE) {
-    const [row] = await supabaseRequest('activity_events', {
-      method: 'POST',
-      body: [
-        {
-          id: event.id,
-          activity_key: activityKey,
-          name: event.name,
-          date_text: event.date,
-          tagline: event.tagline,
-          description: event.description,
-          status: event.status,
-          created_by_name: event.createdBy?.name || '',
-          created_by_email: event.createdBy?.email || '',
-          created_by_phone: event.createdBy?.phone || '',
-        },
-      ],
-    });
-    return sanitizeActivityEventRecord({
-      id: row.id,
-      name: row.name,
-      date: row.date_text,
-      tagline: row.tagline,
-      description: row.description,
-      status: row.status || 'completed',
-      createdAt: row.created_at,
-    });
-  }
-  return withContentLock(async () => {
-    const content = await readContent();
-    content.activityEvents = content.activityEvents || {};
-    content.activityEvents[activityKey] = content.activityEvents[activityKey] || [];
-    content.activityEvents[activityKey].unshift(event);
-    await writeContent(content);
-    return sanitizeActivityEventRecord(event);
-  });
-}
-
-async function deleteActivityEventStore(activityKey, eventId) {
-  if (HAS_SUPABASE) {
-    const rows = await supabaseRequest(
-      `activity_events?activity_key=eq.${encodeURIComponent(activityKey)}&id=eq.${encodeURIComponent(eventId)}`,
-      { method: 'DELETE' }
-    );
-    return Array.isArray(rows) && rows.length > 0;
-  }
-  return withContentLock(async () => {
-    const content = await readContent();
-    content.activityEvents = content.activityEvents || {};
-    const list = content.activityEvents[activityKey] || [];
-    const next = list.filter((e) => e.id !== eventId);
-    if (next.length === list.length) return false;
-    content.activityEvents[activityKey] = next;
-    await writeContent(content);
-    return true;
-  });
-}
-
-async function listCoreTeamStore() {
-  if (HAS_SUPABASE) {
-    const rows = await supabaseRequest('core_team_members?select=*&order=created_at.asc');
-    return rows.map((r) =>
-      sanitizeCoreTeamMemberRecord({
-        id: r.id,
-        name: r.name,
-        role: r.role,
-        year: r.year,
-        branch: r.branch,
-        section: r.section,
-        email: r.email,
-        whatsapp: r.whatsapp,
-        linkedin: r.linkedin,
-        instagram: r.instagram,
-        photoUrl: r.photo_url,
-        createdAt: r.created_at,
-      })
-    );
-  }
-  const content = await readContent();
-  return (content.coreTeam || []).map((member) => sanitizeCoreTeamMemberRecord(member));
-}
-
-const ALLOWED_TEAM_MEMBER_FIELDS = [
-  'id',
-  'name',
-  'role',
-  'position',
-  'bio',
-  'avatar_url',
-  'github_url',
-  'linkedin_url',
-  'email',
-  'joined_at',
-  'order',
-];
-
-function sanitizeCoreTeamMemberRecord(member) {
-  if (!member || typeof member !== 'object') return member;
-  const sanitized = {};
-  for (const field of ALLOWED_TEAM_MEMBER_FIELDS) {
-    if (field in member) sanitized[field] = member[field];
-  }
-  return sanitized;
-}
-
-async function createCoreTeamStore(member) {
-  if (HAS_SUPABASE) {
-    const [row] = await supabaseRequest('core_team_members', {
-      method: 'POST',
-      body: [
-        {
-          name: member.name,
-          role: member.role,
-          year: member.year,
-          branch: member.branch,
-          section: member.section,
-          email: member.email,
-          whatsapp: member.whatsapp,
-          linkedin: member.linkedin,
-          instagram: member.instagram,
-          photo_url: member.photoUrl,
-        },
-      ],
-    });
-    return sanitizeCoreTeamMemberRecord({
-      id: row.id,
-      name: row.name,
-      role: row.role,
-      year: row.year,
-      branch: row.branch,
-      section: row.section,
-      email: row.email,
-      whatsapp: row.whatsapp,
-      linkedin: row.linkedin,
-      instagram: row.instagram,
-      photoUrl: row.photo_url,
-      createdAt: row.created_at,
-    });
-  }
-  return withContentLock(async () => {
-    const content = await readContent();
-    content.coreTeam = content.coreTeam || [];
-    const newMember = {
-      ...member,
-      id: crypto.randomUUID(),
-      createdAt: new Date().toISOString(),
-    };
-    content.coreTeam.push(newMember);
-    await writeContent(content);
-    return sanitizeCoreTeamMemberRecord(newMember);
-  });
-}
-
-async function deleteCoreTeamStore(id) {
-  if (HAS_SUPABASE) {
-    const rows = await supabaseRequest(`core_team_members?id=eq.${encodeURIComponent(id)}`, {
-      method: 'DELETE',
-    });
-    return Array.isArray(rows) && rows.length > 0;
-  }
-  return withContentLock(async () => {
-    const content = await readContent();
-    content.coreTeam = content.coreTeam || [];
-    const before = content.coreTeam.length;
-    content.coreTeam = content.coreTeam.filter((m) => String(m.id) !== String(id));
-    if (content.coreTeam.length === before) return false;
-    await writeContent(content);
-    return true;
-  });
-}
-
-async function appendToSupabaseForms(formType, payload) {
-  if (!HAS_SUPABASE) return false;
-  try {
-    await supabaseRequest('form_submissions', {
-      method: 'POST',
-      body: [
-        {
-          form_type: formType,
-          full_name: toSafeString(payload.fullName, 140),
-          college_email: toSafeString(payload.collegeEmail, 140),
-          whatsapp: toSafeString(payload.whatsapp, 40),
-          payload,
-        },
-      ],
-    });
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-// Constant-time string comparison that does not short-circuit on the first
-// mismatched character. Both operands are encoded to UTF-8 Buffers of equal
-// length before the comparison so response time is independent of how many
-// leading characters match. Returns false immediately if either value is empty,
-// so callers cannot exploit a zero-length buffer edge case.
-function timingSafeStringEqual(a, b) {
-  const sa = String(a ?? '');
-  const sb = String(b ?? '');
-  if (!sa.length || !sb.length) return sa === sb;
-  const ba = Buffer.from(sa, 'utf8');
-  const bb = Buffer.from(sb, 'utf8');
-  // Buffers must be the same byte length for timingSafeEqual. Pad the shorter
-  // one so the comparison always runs the full loop.
-  if (ba.length !== bb.length) {
-    const maxLen = Math.max(ba.length, bb.length);
-    const paddedA = Buffer.alloc(maxLen);
-    const paddedB = Buffer.alloc(maxLen);
-    ba.copy(paddedA);
-    bb.copy(paddedB);
-    // The length mismatch already means they cannot be equal, but we still run
-    // the full comparison so the execution time is data-independent.
-    crypto.timingSafeEqual(paddedA, paddedB);
-    return false;
-  }
-  return crypto.timingSafeEqual(ba, bb);
-}
-
-// Per-IP failed-attempt tracking for the activity-event auth endpoints.
-// Mirrors the passkey lockout pattern used for portfolio mutations below.
-const failedActivityAuthAttempts = new Map();
-const ACTIVITY_AUTH_MAX_ATTEMPTS = 5;
-const ACTIVITY_AUTH_LOCKOUT_MS = 15 * 60 * 1000;
-
-function checkActivityAuthLockout(ip) {
-  const entry = failedActivityAuthAttempts.get(ip);
-  if (!entry) return null;
-  if (Date.now() > entry.lockoutUntil) {
-    failedActivityAuthAttempts.delete(ip);
-    return null;
-  }
-  return entry;
-}
-
-function recordFailedActivityAuth(ip) {
-  const entry = failedActivityAuthAttempts.get(ip) || {
-    count: 0,
-    lockoutUntil: 0,
-  };
-  entry.count += 1;
-  if (entry.count >= ACTIVITY_AUTH_MAX_ATTEMPTS) {
-    entry.lockoutUntil = Date.now() + ACTIVITY_AUTH_LOCKOUT_MS;
-    entry.count = 0;
-  }
-  failedActivityAuthAttempts.set(ip, entry);
-  return entry;
-}
-
-function clearActivityAuthAttempts(ip) {
-  failedActivityAuthAttempts.delete(ip);
-}
 
 // Compliance & Legal Documents (handles both public and admin routes internally)
 app.use('/api/compliance', complianceRouter);
@@ -1271,26 +558,6 @@ app.post('/api/streams/:id/reactions', streamController.addReaction);
 app.get('/api/streams/:id/reactions', streamController.getReactions);
 
 // Public listings
-app.get('/api/content/team', async (req, res) => {
-  try {
-    const rawMembers = await coreTeamService.listMembers();
-    const members = (rawMembers || []).map((m) => {
-      let email = m.email || null;
-      if (email && !email.toLowerCase().endsWith('@glbajajgroup.org')) {
-        email = null;
-      }
-      return {
-        ...m,
-        email,
-        whatsapp: 'https://chat.whatsapp.com/FhpJEaod2g419jFMfqrhGZ',
-      };
-    });
-    return res.json({ members });
-  } catch (e) {
-    return res.status(500).json({ error: e?.message || 'Failed to load core team' });
-  }
-});
-
 // Admin Team Management
 app.get(
   '/api/admin/core-team',
@@ -1313,338 +580,7 @@ app.delete(
   coreTeamController.adminDeleteCoreTeamMember
 );
 
-// Circuit Breaker Admin API
-app.get('/api/admin/circuit-breaker/metrics', adminAuth, async (req, res) => {
-  const metrics = circuitBreakerRegistry.getAllMetrics();
-  return res.json({ circuitBreakers: metrics });
-});
-
-app.post('/api/admin/circuit-breaker/reset/:name', adminAuth, async (req, res) => {
-  const { name } = req.params;
-  const ok = circuitBreakerRegistry.reset(name);
-  if (!ok) {
-    return res.status(404).json({ error: `No circuit breaker found: "${name}"` });
-  }
-  return res.json({ ok: true, message: `Circuit breaker "${name}" reset to CLOSED` });
-});
-
-app.post('/api/admin/circuit-breaker/retry/:name', adminAuth, async (req, res) => {
-  const { name } = req.params;
-  try {
-    const breaker = circuitBreakerRegistry.get(name);
-    if (!breaker) {
-      return res.status(404).json({ error: `No circuit breaker found: "${name}"` });
-    }
-    const result = await breaker.manualRetry();
-    return res.json({ ok: true, state: breaker.state, result });
-  } catch (err) {
-    return res.json({ ok: false, error: err.message });
-  }
-});
-
-// Hard cap on tracked entries. When the limit is reached, the
-// oldest inserted entry is evicted before adding a new one, preventing the
-// Map from growing without bound when an attacker rotates through many
-// distinct usernames from the same or different IP addresses.
-const MAX_PASSKEY_TRACKED_KEYS = 10_000;
-const failedPasskeyAttemptsByIp = new Map();
-const failedPasskeyAttemptsByUsername = new Map();
-
-// Periodic sweep every 30 minutes: remove entries whose lockout period has
-// expired and whose attempt count has already been reset to 0, so they do
-// not accumulate for keys that are never visited again.
-setInterval(
-  () => {
-    const now = Date.now();
-    for (const [key, entry] of failedPasskeyAttemptsByIp) {
-      if (entry.count === 0 && now > entry.lockoutUntil) {
-        failedPasskeyAttemptsByIp.delete(key);
-      }
-    }
-    for (const [key, entry] of failedPasskeyAttemptsByUsername) {
-      if (now > entry.lockoutUntil) {
-        failedPasskeyAttemptsByUsername.delete(key);
-      }
-    }
-  },
-  30 * 60 * 1000
-).unref();
-
-function checkPasskeyLockout(username, ip) {
-  const ipKey = String(ip || 'unknown');
-  const userKey = String(username || '').toLowerCase();
-
-  const ipEntry = failedPasskeyAttemptsByIp.get(ipKey);
-  const userEntry = failedPasskeyAttemptsByUsername.get(userKey);
-
-  const now = Date.now();
-
-  if (ipEntry && ipEntry.lockoutUntil !== 0 && now <= ipEntry.lockoutUntil) {
-    return true;
-  }
-
-  if (userEntry && userEntry.lockoutUntil !== 0 && now <= userEntry.lockoutUntil) {
-    return true;
-  }
-
-  // Cleanup expired entries proactively
-  if (ipEntry && ipEntry.lockoutUntil !== 0 && now > ipEntry.lockoutUntil) {
-    failedPasskeyAttemptsByIp.delete(ipKey);
-  }
-  if (userEntry && userEntry.lockoutUntil !== 0 && now > userEntry.lockoutUntil) {
-    failedPasskeyAttemptsByUsername.delete(userKey);
-  }
-
-  return false;
-}
-
-function recordFailedPasskeyAttempt(username, ip) {
-  const ipKey = String(ip || 'unknown');
-  const userKey = String(username || '').toLowerCase();
-
-  // IP tracking
-  if (
-    !failedPasskeyAttemptsByIp.has(ipKey) &&
-    failedPasskeyAttemptsByIp.size >= MAX_PASSKEY_TRACKED_KEYS
-  ) {
-    failedPasskeyAttemptsByIp.delete(failedPasskeyAttemptsByIp.keys().next().value);
-  }
-  const ipEntry = failedPasskeyAttemptsByIp.get(ipKey) || { count: 0, lockoutUntil: 0 };
-  ipEntry.count += 1;
-  if (ipEntry.count >= 5) {
-    ipEntry.lockoutUntil = Date.now() + 15 * 60 * 1000; // 15 mins
-    ipEntry.count = 0; // Reset count so they need 5 more AFTER lockout to be locked again
-  }
-  failedPasskeyAttemptsByIp.set(ipKey, ipEntry);
-
-  // Username tracking (Exponential backoff)
-  if (
-    !failedPasskeyAttemptsByUsername.has(userKey) &&
-    failedPasskeyAttemptsByUsername.size >= MAX_PASSKEY_TRACKED_KEYS
-  ) {
-    failedPasskeyAttemptsByUsername.delete(failedPasskeyAttemptsByUsername.keys().next().value);
-  }
-  const userEntry = failedPasskeyAttemptsByUsername.get(userKey) || { count: 0, lockoutUntil: 0 };
-  userEntry.count += 1;
-  if (userEntry.count >= 5) {
-    // 5 attempts = 1 min, 6 = 2 mins, 7 = 4 mins, 8 = 8 mins, 9+ = 15 mins
-    const factor = Math.pow(2, Math.max(0, userEntry.count - 5));
-    const delayMinutes = Math.min(15, factor);
-    userEntry.lockoutUntil = Date.now() + delayMinutes * 60 * 1000;
-  }
-  failedPasskeyAttemptsByUsername.set(userKey, userEntry);
-
-  return { ipEntry, userEntry };
-}
-
-function clearPasskeyAttempts(username, ip) {
-  const ipKey = String(ip || 'unknown');
-  const userKey = String(username || '').toLowerCase();
-
-  failedPasskeyAttemptsByIp.delete(ipKey);
-  failedPasskeyAttemptsByUsername.delete(userKey);
-}
-
-app.get('/api/notifications', async (req, res) => {
-  try {
-    const userId = req.query.userId || 'global';
-
-    if (userId !== 'global') {
-      let authenticated = false;
-
-      // 1. Try Student Auth
-      let token = null;
-      const authHeader = req.headers.authorization;
-      if (authHeader && authHeader.startsWith('Bearer ')) {
-        token = authHeader.slice(7).trim();
-      }
-      if (!token && req.cookies?.ns_student_token) {
-        token = req.cookies.ns_student_token;
-      }
-      if (token) {
-        const payload = studentAuthService.verifyToken(token);
-        if (payload && (payload.sub === userId || payload.id === userId)) {
-          authenticated = true;
-        }
-      }
-
-      // 2. Try Admin Auth
-      if (!authenticated) {
-        let adminToken = null;
-        if (authHeader && authHeader.startsWith('Bearer ')) {
-          adminToken = authHeader.slice(7).trim();
-        }
-        if (!adminToken && req.cookies?.ns_admin_token) {
-          adminToken = req.cookies.ns_admin_token;
-        }
-        if (adminToken) {
-          const { getAdminSession } = await import('./repositories/adminSessionsRepository.js');
-          const session = await getAdminSession(adminToken);
-          if (session) {
-            authenticated = true;
-          }
-        }
-      }
-
-      if (!authenticated) {
-        return res.status(401).json({ error: 'Unauthorized to view these notifications' });
-      }
-    }
-
-    const offset = parseInt(req.query.offset, 10) || 0;
-    const limit = Math.min(parseInt(req.query.limit, 10) || 100, 500);
-    const list = await notificationsService.getNotifications(userId, offset, limit);
-    return res.json({ notifications: list });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-function requireNotificationPrefAuth(req, res, next) {
-  adminAuthMiddleware.requireAdmin(req, res, (err) => {
-    if (!err && req.adminSession) {
-      return next();
-    }
-    requireStudentAuth(req, res, (err2) => {
-      if (err2 || !req.studentUser) {
-        return res.status(401).json({ error: 'Unauthorized: Authentication required' });
-      }
-      const userId =
-        req.method === 'GET' ? req.query.userId || 'global' : req.body.userId || 'global';
-      if (req.studentUser.sub === userId || req.studentUser.id === userId) {
-        return next();
-      }
-      return res
-        .status(403)
-        .json({ error: "Forbidden: You cannot access or modify other users' preferences" });
-    });
-  });
-}
-
-// Notification Preferences
-app.get('/api/notifications/preferences', requireNotificationPrefAuth, async (req, res) => {
-  try {
-    const userId = req.query.userId || 'global';
-    const prefs = await notificationPreferencesRepository.list(userId);
-    return res.json({ preferences: prefs });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-app.put('/api/notifications/preferences', requireNotificationPrefAuth, async (req, res) => {
-  try {
-    const userId = req.body.userId || 'global';
-    const { category, email, push, in_app, sms, frequency, quiet_start, quiet_end, dnd } = req.body;
-    if (!category) return res.status(400).json({ error: 'category is required' });
-    const pref = await notificationPreferencesRepository.set(userId, category, {
-      email,
-      push,
-      in_app,
-      sms,
-      frequency,
-      quiet_start,
-      quiet_end,
-      dnd,
-    });
-    return res.json({ preference: pref });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-app.put('/api/notifications/preferences/bulk', requireNotificationPrefAuth, async (req, res) => {
-  try {
-    const userId = req.body.userId || 'global';
-    const { preferences } = req.body;
-    if (!Array.isArray(preferences) || !preferences.length) {
-      return res.status(400).json({ error: 'preferences array is required' });
-    }
-    const results = await notificationPreferencesRepository.setBulk(userId, preferences);
-    return res.json({ preferences: results });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-// Notification analytics (lightweight collector)
-app.post('/api/notifications/analytics', async (req, res) => {
-  try {
-    const event = req.body || {};
-    // Minimal validation — in future route can forward to analytics pipeline
-    console.log('[notification-analytics]', event.type || 'unknown', event);
-    return res.json({ ok: true });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-app.put('/api/portfolio', portfolioRateLimiter, async (req, res) => {
-  try {
-    const body = req.body || {};
-    const ip = req.ip || req.headers['x-forwarded-for'] || 'unknown';
-
-    // 1. Validate credentials up front.  Anything below this point
-    //    trusts the username + passkey pair.
-    const credentials = portfolioPutSchema.safeParse({
-      username: body.username,
-      passkey: body.passkey,
-    });
-    if (!credentials.success) {
-      const firstIssue = credentials.error.issues[0];
-      return res.status(400).json({ error: firstIssue?.message || 'Invalid request body' });
-    }
-    const { username, passkey } = credentials.data;
-
-    // 2. Validate the content body.  This rejects XSS payloads such
-    //    as javascript: URLs and unknown protocol schemes before
-    //    the data ever reaches the repository.  The repository
-    //    re-sanitizes as defense-in-depth.
-    const content = portfolioContentSchema.safeParse(body);
-    if (!content.success) {
-      const firstIssue = content.error.issues[0];
-      return res.status(400).json({
-        error:
-          `Invalid portfolio content: ${firstIssue?.path?.join('.') || ''} ${firstIssue?.message || ''}`.trim(),
-      });
-    }
-
-    const existingPortfolio = await portfolioRepository.getByUsername(username);
-    const isNewRegistration = !existingPortfolio;
-
-    const lockout = checkPasskeyLockout(username, ip);
-    if (lockout) {
-      return res.status(429).json({
-        error: 'Too many failed passkey attempts. Please try again later.',
-      });
-    }
-
-    const isAuthorized = await portfolioRepository.verifyPasskey(username, passkey, {
-      allowNew: isNewRegistration,
-    });
-    if (!isAuthorized) {
-      recordFailedPasskeyAttempt(username, ip);
-      return res.status(401).json({ error: 'Incorrect passkey for this username' });
-    }
-
-    clearPasskeyAttempts(username, ip);
-
-    const saved = await portfolioRepository.createOrUpdate({
-      ...content.data,
-      username,
-      passkey,
-    });
-    return res.json({ ok: true, portfolio: saved });
-  } catch (err) {
-    if (err.code === '23505') {
-      return res
-        .status(409)
-        .json({ error: 'Username already exists. Another request may have just created it.' });
-    }
-    console.error('Error saving portfolio:', err);
-    return res.status(500).json({ error: err.message || 'Internal server error' });
-  }
-});
+app.use('/api/admin/circuit-breaker', circuitBreakerRouter);
 
 // ── Forum / Q&A ──
 app.get('/api/forum/categories', forumController.listCategories);
@@ -1664,19 +600,6 @@ app.patch('/api/admin/forum/threads/:id/moderate', adminAuth, forumController.mo
 app.patch('/api/admin/forum/replies/:replyId/moderate', adminAuth, forumController.moderateReply);
 app.get('/api/admin/forum/threads', adminAuth, forumController.adminListThreads);
 
-function requireMentorshipAuth(req, res, next) {
-  adminAuthMiddleware.requireAdmin(req, res, (err) => {
-    if (!err && req.adminSession) {
-      return next();
-    }
-    requireStudentAuth(req, res, (err2) => {
-      if (!err2 && req.studentUser) {
-        return next();
-      }
-      return res.status(401).json({ error: 'Unauthorized: Authentication required' });
-    });
-  });
-}
 
 // ── Mentorship & Buddy System ──
 app.get('/api/mentorship/mentors', mentorshipController.listMentors);
