+++ server/routes/coreTeam.js
@@ -15,25 +15,7 @@ const adminAuth = adminAuthMiddleware.requireAdmin;
  * GET /api/content/team — Public core team listing.
  * Filters out non-@glbajajgroup.org emails for privacy.
  */
-router.get('/api/content/team', async (req, res) => {
-  try {
-    const rawMembers = await coreTeamService.listMembers();
-    const members = (rawMembers || []).map((m) => {
-      let email = m.email || null;
-      if (email && !email.toLowerCase().endsWith('@glbajajgroup.org')) {
-        email = null;
-      }
-      return {
-        ...m,
-        email,
-        whatsapp: 'https://chat.whatsapp.com/FhpJEaod2g419jFMfqrhGZ',
-      };
-    });
-    return res.json({ members });
-  } catch (e) {
-    return res.status(500).json({ error: e?.message || 'Failed to load core team' });
-  }
-});
+router.get('/api/content/team', coreTeamController.publicListMembers);
 
 /**
  * GET /api/admin/core-team — List all core team members (admin).
